requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator
requests
razorpay
stripe
PyJWT
httpx
psycopg2-binary
cors-fastapi
twilio
Pillow
